x32: got the syscall semantics right - all tests pass
the actual syscall functions need to take long long arguments, but return
long. the syscall macros must cast the arguments to unsigned long, other-
wise pointers will get sign-extended to 64bit and cause crashes.

+++ arch/x32/syscall_arch.h
@@ -1,50 +1,50 @@
 #define __SYSCALL_LL_E(x) (x)
 #define __SYSCALL_LL_O(x) (x)
 
-static __inline long long __syscall0(long long n)
+static __inline long __syscall0(long long n)
 {
-	unsigned long long ret;
+	unsigned long ret;
 	__asm__ __volatile__ ("syscall" : "=a"(ret) : "a"(n) : "rcx", "r11", "memory");
 	return ret;
 }
 
-static __inline long long __syscall1(long long n, long long a1)
+static __inline long __syscall1(long long n, long long a1)
 {
-	unsigned long long ret;
+	unsigned long ret;
 	__asm__ __volatile__ ("syscall" : "=a"(ret) : "a"(n), "D"(a1) : "rcx", "r11", "memory");
 	return ret;
 }
 
-static __inline long long __syscall2(long long n, long long a1, long long a2)
+static __inline long __syscall2(long long n, long long a1, long long a2)
 {
-	unsigned long long ret;
+	unsigned long ret;
 	__asm__ __volatile__ ("syscall" : "=a"(ret) : "a"(n), "D"(a1), "S"(a2)
 						  : "rcx", "r11", "memory");
 	return ret;
 }
 
-static __inline long long __syscall3(long long n, long long a1, long long a2, long long a3)
+static __inline long __syscall3(long long n, long long a1, long long a2, long long a3)
 {
-	unsigned long long ret;
+	unsigned long ret;
 	__asm__ __volatile__ ("syscall" : "=a"(ret) : "a"(n), "D"(a1), "S"(a2),
 						  "d"(a3) : "rcx", "r11", "memory");
 	return ret;
 }
 
-static __inline long long __syscall4(long long n, long long a1, long long a2, long long a3,
+static __inline long __syscall4(long long n, long long a1, long long a2, long long a3,
                                      long long a4)
 {
-	unsigned long long ret;
+	unsigned long ret;
 	register long long r10 __asm__("r10") = a4;
 	__asm__ __volatile__ ("syscall" : "=a"(ret) : "a"(n), "D"(a1), "S"(a2),
 						  "d"(a3), "r"(r10): "rcx", "r11", "memory");
 	return ret;
 }
 
-static __inline long long __syscall5(long long n, long long a1, long long a2, long long a3,
+static __inline long __syscall5(long long n, long long a1, long long a2, long long a3,
                                      long long a4, long long a5)
 {
-	unsigned long long ret;
+	unsigned long ret;
 	register long long r10 __asm__("r10") = a4;
 	register long long r8 __asm__("r8") = a5;
 	__asm__ __volatile__ ("syscall" : "=a"(ret) : "a"(n), "D"(a1), "S"(a2),
@@ -52,10 +52,10 @@ static __inline long long __syscall5(long long n, long long a1, long long a2, lo
 	return ret;
 }
 
-static __inline long long __syscall6(long long n, long long a1, long long a2, long long a3,
+static __inline long __syscall6(long long n, long long a1, long long a2, long long a3,
                                      long long a4, long long a5, long long a6)
 {
-	unsigned long long ret;
+	unsigned long ret;
 	register long long r10 __asm__("r10") = a4;
 	register long long r8 __asm__("r8") = a5;
 	register long long r9 __asm__("r9") = a6;
