+++ src/internal/syscall_ret.c
@@ -1,6 +1,6 @@
 #include <errno.h>
 
-long long __syscall_ret(unsigned long long r)
+long __syscall_ret(unsigned long r)
 {
 	if (r > -4096UL) {
 		errno = -r;
